refactor(phase-5d): extract wp.desktop.* facade to src/api/facade.ts
Pulls the ~280-LOC desktopApi literal and the
RESERVED_NAMESPACE_KEYS allowlist out of init() into a dedicated
facade module. desktop.ts is now 2,623 LOC (down from 3,034 at
start of phase 5; 1,072 LOC moved across all 5d sub-commits).

- src/api/facade.ts (552 LOC) — owns:
    * `BuildPublicApiDeps` interface listing every dependency
      init() needs to produce in order to build the API
    * `buildPublicApi(deps): WpDesktopPublicApi` — pure builder,
      no side effects on window.wp.desktop
    * `installPublicApi(api)` — the merge-onto-shim assignment
      (Object.assign so the early-shim's whenReady queue
      doesn't get orphaned)
    * `RESERVED_NAMESPACE_KEYS` — the source-of-truth allowlist
      that registerNamespace consults; lives next to the
      assembly so it stays in sync
- desktop.ts: drops the inline literal + the allowlist + ~30
  imports that became facade-only (registerWindowTheme/Control/
  Slot/Chrome + unregister/list, broadcast/subscribe, all PWA
  notify/install/state helpers, devtools, createSharedStore,
  presenceApi, activity, heartbeat, listDockRailRenderers and
  friends, listCommands/listPalettes, etc.). The init() block
  now reads as 22 lines: build a deps bag, call buildPublicApi,
  call installPublicApi.

Behavior unchanged. Plugin authors see the same 60+ method
surface on `wp.desktop.*`; tests exercising the public API
continue to pass.

All green: lint, tsc --noEmit, vitest (106 files / 936 tests),
npm run build.
